+++ src/playback/playhead.rs
@@ -7,28 +7,28 @@ pub struct Playhead {
     /// Used to track transitions relating to time.
     pub(crate) first_render: Option<Instant>,
     /// The actual frame being rendered
-    pub(crate) frame: f32,
+    pub(crate) frame: f64,
     /// Used to track intermission.
     pub(crate) intermission: Option<Timer>,
     /// Used to count loops for loop behavior.
     pub(crate) loops_completed: usize,
     /// Used by play mode to track current direction. Only set to -1.0
     /// (reverse) or 1.0 (normal).
-    pub(crate) playmode_dir: f32,
+    pub(crate) playmode_dir: f64,
 }
 
 impl Playhead {
     /// Get the current playhead frame
-    pub fn frame(&self) -> f32 {
+    pub fn frame(&self) -> f64 {
         self.frame
     }
 
     /// Seek to a given frame
-    pub fn seek(&mut self, frame: f32) {
+    pub fn seek(&mut self, frame: f64) {
         self.frame = frame;
     }
 
-    pub(crate) fn new(frame: f32) -> Self {
+    pub(crate) fn new(frame: f64) -> Self {
         Self {
             frame,
             first_render: None,
